Project: Add Navgraph
with Auth Activity

+++ data/src/main/java/com/example/data/dto/User.kt
@@ -1,11 +1,10 @@
 package com.example.data.dto
 
 import com.google.gson.annotations.SerializedName
-
 data class User(
     @SerializedName("createdAt") var createdAt: String? = null,
     @SerializedName("updatedAt") var updatedAt: String? = null,
-    @SerializedName("id") var id: Int? = null,
+    @SerializedName("id") var id: Long? = null,
     @SerializedName("email") var email: String? = null,
     @SerializedName("password") var password: String? = null,
     @SerializedName("name") var name: String? = null,
